[avalanche] Allow for yes-voting several winners
Summary:
This updates the getstakingreward and setstakingreward RPC so that an array of winners can be manually set. This makes it possible to allow 2 winners manually if there is a doubt about which is the correct one (e.g. if 2 nodes disagree).

Also use this to allow for more winners if we are missing connection to the selected proof, and leverage the remote proofs to determine if a proof looks abnormal.
Mining remains unchanged, only the node vote is affected.

Depends on D15685.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15686