[Cashtab] Use in-node chronik
Summary:
Use in-node chronik to power Cashtab

Because in-node chronik returns `utxos` and `tx` objects in a different shape, many existing mocks are updated in this diff. For the most part this was scripted (i.e., call for this tx from an in-node chronik instance and use that output).

This diff impplements in-node chronik-client with minimal behavior change to Cashtab. In-node chronik makes future improvements possible. These will be in later diffs (e.g. we can get avalanche finalization status on txs from the websocket, we can deprecate SLPDB in airdrops, we can support more tokens than slpv1, etc).

Test Plan:
`npm test`

This version is mock prod-deployed at https://cashtab-local-dev.netlify.app/

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: tobias_ruck, emack

Differential Revision: https://reviews.bitcoinabc.org/D15693