[Cashtab] Clean up how fiat price for min token creation is displayed
Summary:
`settings` will never not be defined here, so stop checking. Do not render a fiatPrice if you do not have it yet.

Change here because the app used to always load a USD fiat price immediately. Now the app waits to load the fiat price in the user's selected currency. It still renders, but it's not straightforward to get react testing library to wait for it. So, we test that it works with no fiat price.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15703