[Cashtab] Render onboarding screen if user does not have a wallet
Summary:
For the special case of a user with no wallet in storage, i.e. a new user loading Cashtab for the first time -- show the onboarding screen.

This should be standardized for all routes. Right now, it has custom implementations on some routes that arose piecemeal in Cashtab dev history.

Because this diff changes the loading behavior of Cashtab, a number of integration tests are updated.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15711