[Cashtab] Change wallet from dropdown menu in header
Summary:
T2871 T3224

It is inconvenient UX to require the user to go to Configure in order to change wallet. Now that wallets are better organized, we can do this with a standardized method from other screens.

Test Plan:
`npm test`

check on test site: https://cashtab-local-dev.netlify.app/#/wallet

change wallets a lot from the dropdown, try on mobile

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15764