[chronik-client] Add timeout failure to ws connections
Summary: In local testing, occasionally a ws will hang and fail to connect. In this case, we should try the next ws.

Test Plan:
`npm test`

I have used this while locally testing Cashtab and can confirm the timeout failures will cause failover proxy to use the next ws. We may want to tweak the timeout depending on behavior we see with apps.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15771