cleanup more unused code after removal of chained-tx logic (TestLockPointValidity)
Summary:
`TestLockPointValidity` is unused since D13133. Remove it, and remove include for `GetConsensus` that is no longer used.

`LockPoint::maxInputBlock` is no longer used by anything, `TestLockPointValidity` was the only user. We no longer need to compute it in validation and we can remove it from the struct.
As a result we longer need to include chain.h (CBlockIndex).

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15788