[avalanche] Account for network propagation time for staking rewards
Summary: There is a delay before a proof can be selected for staking rewards, which can cause discrepeancy due to the proof propagation/registration time. To avoid this we allow for an additional winner if the proof is in a recent registration time range. This does not change the winner selected by the miner.

Test Plan:
  ninja check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D15791