[CI] Make sure to upload the artifacts upon failure
Summary: The builds are running in containers and the permissions for the docker user are not the permissions for the host user. Relaxing the permissions for the artifacts make sure the CI can create an upload the archive from the container volume files.

Test Plan: See the previous iteration that created test failures on purpose, check the artifacts are uploaded as expected.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15793