get config via chainstatemanager when possible (1/n)
Summary:
This removes the need for passing config as a function argument through the whole validation stack, when a reference is already available via `ChainstateManager.options`.

In this diff we apply this idea to `AcceptToMemoryPool`, `ProcessNewPackages` and `ConnectTip` (and to the whole stake of functions that needed a `Config` for the only purpose calling these 3)

There are more places where this can be applied, but I noticed the diff is starting to get big, so let's do the rest in separate diffs.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15796