[Cashtab] Parse mint txs
Summary:
Parse slpv1 mint txs

Note: this is not a great implementation. It follows the pattern of existing tx history parsing. However, this is a bad pattern. I am working on a refactor but that will be a multi-part diff that will end up replacing the whole Tx component.

Since the component will be replaced anyway, imo it is worth parsing these txs right now, even if the implementation is not great. Alternative is these txs are parsed as -0 XEC txs, confusing and wrong.

Test Plan:
`npm test`

{F7455739}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15837