[avalanche] Fix a race in abc_p2p_avalanche_remote_proofs.py
Summary: The compact proofs are sent periodically so it's possible for the trigger_avaproofs() function to return on a previously sent message and cause the test to fail. This diff addresses this by using a retry loop and checking the last_update time to make sure we're looking at the correct message.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_avalanche_remote_proofs

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15843