[Cashtab] Patch bug in disable send logic
Summary:
The send button will not be disabled if user has valid inputs for address and amount in send-to-one mode. The tx will not work and no error msg is thrown.

Correct logic so that the send button is disabled if the user is in send-to-many mode and has no send-to-many input.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15845