[electrum] add the option to always add new addresses to wallet when scanning beyond gap
Summary:
Users have requested a better solution for generating more addresses without having to use console command.
This adds a checkbox to the "Scan More Addresses" tool. Users can optionnaly check it to save all the newly scanned addresses to the wallet, no matter if any tx history was found for any following address.

Note that as a result of this the wallet will monitor more unused addresses than the usual 20 (default gap limit), so it is not something I recommend doing for wallets with already a large history.

Depends on D15848

Test Plan:
use the Scan More Addresses tool  with the new checkbox toggled. Make sure the expected number of new addresses is added to the Address tab.
Try scanning for only receive addresses, only change addresses, and both.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15849