[Cashtab] Improvements to send to many switch
Summary:
Now that switch is becoming a normal part of the UI and not a one-off thing, keeping the send-to-many switch aligned right looks off. Move it to the left.

Also fix its data-test-id which is super confusing when reviewing unit tests. Was set that way as I thought I would be able to easily get the component using a name query, but alas this was not the case.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15862