[Cashtab] Move App to components/App and remove lingering relative imports
Summary:
In theory, it makes sense that `App.js` is at the top level of `src/components`, with other components one level deeper, as `App.js` is the parent component.

In practice, it is always confusing to find the tests for App.js, as its test directory appears at the top of `src/components`, while its `fixtures` dir is in the middle, then, finally, `App.js` is below the other directories.

Going forward, may want to reconsider how tests are organized in general. We do not need to use the `__tests__` directories in the latest version of jest (which we use now). This though would be another diff.

For now, make the organization make more sense based on practical experience of managing the repo.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15865