[Cashtab] Add op return raw field to send xec screen
Summary:
T3460

Add support for op_return_raw input to Cashtab's SendXec UI.

Currently, this can be set by `bip21`. However, it is confusing and scary when it is set this way, because there is no parsing of the `op_return_raw` string -- so the user does not know what the tx is.

Add parsing and render a preview window of the parsed op_return_raw. Support all protocols known to Cashtab.

Instead of the kind of useless "bip 21 alert" divs -- we now just configure the form the way a user would do manually and disable the inputs.

The "webapp tx request" is preserved, for now. Will clean this up in a separate diff.

Test Plan:
`npm test`

This is live at `https://cashtab-local-dev.netlify.app/`

Test with extension

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15874