[chronik] Don't show "Stopping Chronik..." in the log when chronik is not enabled
Summary: The chronik::Stop() function is called unconditionally if chronik is built-in, even if it is not enabled via the node option which is confusing.

Test Plan:
Build with Chronik, then:
  ./src/bitcoind
After a few seconds, kill bitcoind with CTRL+C and check the log doesn't contain any chronik related line.

Reviewers: #bitcoin_abc, tobias_ruck

Reviewed By: #bitcoin_abc, tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D15878