[Cashtab] Use cursor not-allowed when select menu is disabled
Summary:
This menu can be disabled when the price API is down or when a tx is populated by a URL / webapp.

Show a 'no' sign for the cursor on hover.

Test Plan:
`npm start`, navigate to `http://localhost:3000/#/send?bip21=ecash:qp89xgjhcqdnzzemts0aj378nfe2mhu9yvxj9nhgg6?amount=17`, mouse over, cursor is not allowed

{F7468796}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15883