[Cashtab] Move contact lists to own screen
Summary:
T2207

Move contact list to own screen. Rebuild to use better styles and code.

Ensure same features are met by upgrading integration tests.

Once again, way too much time spent on Contacts. But, has to be done if we are to maintain the feature. Some upside here is that we are getting lots of lessons learned and best practices for managing formData and modals. Compare the implementation in Configure with the new Contacts implementation.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15884