[electrum] build: fix incompatible dependencies
Summary:
In D15870 we bumped the pinpointed `pyopenssl` version to 24.1.0, which introduced a dependency on `cryptography>42`. However the currently pinpointed version of `dnspython[dnssec]` had a incompatible limit on `cryptography` version (`<40`).
Note that prior to D15870 we already had an issue because of pinpointing `cryptography==41.x`, but the builds worked anyway with just warnings. But now the MacOS build is broken.

Bump  `dnspython[dnssec]` to the latest release, which dropped the upper limit for `cryptography` and definitely supports `==42`

Test Plan: Build all the release binaries and test them.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15885