[Cashtab] Remove unused validation function
Summary:
In D15884, we tightened validation for contact list (only accept ecash: prefixed addresses). So, we stopped using this function. If we start accepting aliases, we need to convert them to addresses before storing them to keep app functionality straightforward.

For now, this is not used anywhere, so get rid of it.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15919