[explorer] Switch to new chronik
Summary:
This diff contains the last steps that cannot be split apart to migrate the explorer to the in-node chronik:
 - Update the pinned version of the dependencies so it supports the in-node chronik
 - Update block txs retrieval via the new block-txs endpoint
 - Accommodate the new API/Protobuf

The token parsing is assuming a single action per token and only the first one (if any) is displayed. This can be improved in follow ups.
This diff is only targeting feature parity with the NNG version.

Depends on D15962.

Test Plan:
  cargo run

Reviewers: #bitcoin_abc, PiRK, tobias_ruck

Reviewed By: #bitcoin_abc, PiRK, tobias_ruck

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D15930