[Cashtab] Remove data-testid from everything but Select menus
Summary:
T2207

data-testid should not be used, we should access elements the way a user would interact with the app

for now we still use it in the select menus. can figure that one out and implement a one-off, or tolerate this minor remaining presence.

Test Plan:
`npm test`

this diff is deployed at https://cashtab-local-dev.netlify.app/

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15942