[Cashtab] Standardize svg icon buttons
Summary:
T3528

Starting to implement NFTs in Cashtab -- requires a number of UI improvements. After the recent UI overhaul I would prefer to avoid tech debt as much as possible. So, there are some basic UI improvements to get in before building the NFT interface.

Here we standardize the icon button, a component that is re-used on different screens with different implementations.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15951