[Cashtab] Better switch mgmt in send token
Summary:
T3528

For send token, we only want one token feature enabled at a time. We will be adding more token features for different token types, and it is already a bit of a mess to manually turn off "all the other switches" every time a switch comes on.

Standardize the approach so it will be easier to add NFT features.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15953