[Cashtab] After a token is created, automatically navigate to its actions screen
Summary:
T3528

The way Cashtab works right now, after you create a token, you must find it from the mystery list. Also you must know where the mystery list of tokens is (you click the mystery token icon, duh). Also it is of course not in alphabetical order.

The mystery list will be improved separately. Regardless, it makes sense for the user to be forwarded to the token they just created.

Note: we have a race condition in that Cashtab will update the wallet and navigate to the token page before the token server has resized its icon. I do not think it is wise to add a `setTimeout` to (maybe) handle this. We can optmize this later if people are confused.

Test Plan:
`npm test`, can try live at https://cashtab-local-dev.netlify.app/ where a version with console.log statements is up (and the token server will take icons from this url)

{F7488333}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15955