[Cashtab] No scrolling of no content below fold
Summary: Some screens fit all content above the fold. When this happens, we do not want the user to be able to scroll down and see a lot of empty space. Want to emulate a mobile app experience.

Test Plan:
`npm test`

This diff is live at https://cashtab-local-dev.netlify.app/

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D15972