[Cashtab] Improve focus-visible for wallet dropdown menu
Summary:
Tempting to just lose this as it does not look great, but for users who want to select a wallet with the keyboard need some kind of indicator it is selected. The browser default (a black oval) is not really in line with Cashtab's styling and also gets a bit too close to the balance row beneath.

T3542

Instead, underline the text.

Test Plan: `npm test`, `npm start`, click the dropdown menu, after clicking it is underlined and looks ok.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15982