[Cashtab] Do not render reply option to eCash chat
Summary:
T3540

Cashtab does not support replying to an ecash chat with an on-spec ecash chat. It does not make sense for 'reply' to take the user to sending a Cashtab Msg to the eCash chat sender (current behavior).

So, do not render a reply button for received eCash chat.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15985