[Cashtab] Add animation for avalanche finalized txs
Summary:
T3461

Animated check mark when a tx is finalized. For now, this only happens when a block is finalized.

Test Plan:
`npm test`, review aesthetics

{F7496440}
{F7496543}

this diff is live at https://cashtab-local-dev.netlify.app/

on the test site (or with `npm start`), send yourself a tx. observe it shows inline loader in history and "confirming" + inline loader in tx panel when expanded. Wait for tx to confirm. Observe that the tx updates automatically to finalized (on the dev site, there are also dev logs for this).

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15987