[Cashtab] Better hamburger menu
Summary:
We still want to migrate to horizontal scrolling nav buttons or mb a sidebar, but we are also still adding screens so it could be awhile.

In the meantime, the hamburger menu could use an update.

Make it bigger, get rid of empty space to the left of and between some buttons, use px instead of rem since we are not really font-size dependent for these dimensions, do not use float which does not work in display flex

Test Plan:
`npm test`, `npm start` and aesthetic judgments

before
{F7497305}

after
{F7497304}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D15995