[Cashtab] Start reward countdown automatically after claiming rewards
Summary:
After claiming a reward, eligibility resets (per address). Right now, Cashtab does not automatically make an API request, so the button remains enabled (and will throw an error if the user clicks it).

Better to have the button automatically change to the countdown after a claim.

Test Plan: `npm start`, claim a reward with an eligibile wallet, notice the countdown appears

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16014