[Cashtab] Word-break for validation msgs
Summary:
T3553

Long validation msgs can be cut-off as a long string early on will bump to the second line.

Add word-break: break-all to reduce this issue.

Note: The issue is not totally "solved," but this is a Cashtab design decision (validation error msgs are restricted to 2 lines. So, it is possible to generate one longer than 2 lines and see a cutoff).

Test Plan:
`npm test`

Before:
{F7505653}

solves problem for user-reported (and expected common) case:
{F7505652}

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D16031