[Cashtab] Support for creating NFT Collections
Summary:
Update the create token form to support creating NFT1 Parent tokens, which we call "NFT Collections" in Cashtab.

We take a sha256 hash of the user's raw file uploaded for the image and set this as the token document hash field.

We do not support minting NFTs yet, but display an Info notice on the Token page explaining this feature is coming soon.

Test Plan:
`npm test`

This diff is live at https://cashtab-local-dev.netlify.app/ if you want to test creating an NFT collection with an icon (token-server still blocks doing this from localhost)

To test the hash:

- Pick a jpg or png
- Get its sha256 hash from your local machine, e.g. `sha256sum myjpg.jpg`
- Note this
- Upload the jpg on the create token form with "NFT Collection" selected
- Confirm the token document hash in the new input field matches what you got on your local machine

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16032